feat: 릴리스정보 엔티티, 응답모델 구조 변경 - ConfirmOrder 엔티티  ManyToOne 연관관계로 갖고있도록 수정 - Claim 엔티티 OneToOne 연관관계로 갖고있도록 추가

+++ src/main/kotlin/com/smartfoodnet/apiclient/response/NosnosDeliveryAgencyInfoModel.kt
@@ -5,6 +5,6 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming
 
 @JsonNaming(value = PropertyNamingStrategies.SnakeCaseStrategy::class)
 data class NosnosDeliveryAgencyInfoModel(
-    val deliveryAgencyId: Int? = null,
+    val deliveryAgencyId: Long? = null,
     val deliveryAgencyName: String? = null,
 )
