+++ src/main/kotlin/com/smartfoodnet/fninventory/stock/StockScheduledService.kt
@@ -74,7 +74,7 @@ class StockScheduledService(
     fun deleteDuplicateSummaries(dateRange: Long) {
         val effectiveDates = mutableListOf<LocalDate>()
         val today = LocalDate.now()
-        for (i in 1 .. dateRange) {
+        for (i in 1..dateRange) {
             effectiveDates.add(today.minusDays(i))
         }
 
@@ -245,7 +245,7 @@ class StockScheduledService(
 
     private fun calculateBestBefore(
         expirationDate: LocalDateTime?,
-        manufacturedBefore: Int?
+        manufacturedBefore: Long?
     ): Float {
         if (expirationDate == null || manufacturedBefore == null) {
             return -1F
